fix(mobile): cycle dependencies
